+++ backend/src/main/java/si/feri/ris/kirbis/todo/repositories/TasklistRepository.java
@@ -3,5 +3,5 @@ package si.feri.ris.kirbis.todo.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import si.feri.ris.kirbis.todo.entities.Tasklist;
 
-public interface TasklistRepository extends JpaRepository<Tasklist, Long> {
+public interface TasklistRepository extends JpaRepository<Tasklist, Integer> {
 }
